pyproject.toml
[tool.poetry]
name = "web_condenser_ai"
version = "0.1.0"
description = ""
authors = ["thisishugow <59921505+thisishugow@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "web_condenser_ai"}]

[tool.poetry.dependencies]
python = ">3.9.7,<3.12"
langchain = "^0.2.2"
python-dotenv = "^1.0.1"
langchain-google-genai = "^1.0.5"
google-generativeai = "^0.5.2"
selenium = "^4.21.0"
unstructured = "^0.14.4"
langchain-community = "^0.2.3"
langchain-openai = "^0.1.8"
streamlit = "^1.35.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
